Sites_Management_API/SitesManagementAPI/requirements.txt
asgiref==3.5.2
django==4.1.2
django-rest-framework==0.1.0
pytz==2021.1
sqlparse==0.4.3
